# docs/swagger.yaml
basePath: /api/v1
definitions:
  api.APIError:
    properties:
      message:
        type: string
      path:
        type: string
      status:
        type: integer
      timestamp:
        type: integer
    type: object
  model.Post:
    properties:
      author:
        type: string
      content:
        type: string
      date:
        type: string
      id:
        type: integer
      title:
        type: string
    required:
    - author
    - content
    - id
    - title
    type: object
  model.User:
    properties:
      name:
        type: string
      password:
        type: string
    required:
    - name
    - password
    type: object
host: localhost:8080
info:
  contact: {}
  title: Golang CHATTING REST API
  version: "1.0"
paths:
  /join:
    post:
      consumes:
      - application/json
      description: user Save
      parameters:
      - description: user info
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/model.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.User'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.APIError'
      summary: Post User
      tags:
      - users
  /login:
    post:
      consumes:
      - application/json
      description: User login
      parameters:
      - description: User Info
        in: body
        name: name
        required: true
        schema:
          $ref: '#/definitions/model.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.User'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.APIError'
      summary: User Login
      tags:
      - users
  /posts:
    get:
      consumes:
      - application/json
      description: post get All
      parameters:
      - description: access_token
        in: header
        name: access_token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Post'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.APIError'
      summary: Get all Posts
      tags:
      - posts
    post:
      consumes:
      - application/json
      description: save post
      parameters:
      - description: post Info
        in: body
        name: post
        required: true
        schema:
          $ref: '#/definitions/model.Post'
      - description: access_token
        in: header
        name: access_token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Post'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.APIError'
      summary: Save Post
      tags:
      - posts
  /posts/{id}:
    delete:
      consumes:
      - application/json
      description: Delete post
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: integer
      - description: access_token
        in: header
        name: access_token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Post'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.APIError'
      summary: delete post
      tags:
      - posts
    get:
      consumes:
      - application/json
      description: Get one post
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: integer
      - description: access_token
        in: header
        name: access_token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Post'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.APIError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.APIError'
      summary: Get One Post
      tags:
      - posts
    put:
      consumes:
      - application/json
      description: Update post
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: integer
      - description: post Info
        in: body
        name: post
        required: true
        schema:
          $ref: '#/definitions/model.Post'
      - description: access_token
        in: header
        name: access_token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Post'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.APIError'
      summary: Update Post
      tags:
      - posts
swagger: "2.0"